crank de 2 bytes

+++ main/ble.h
@@ -16,4 +16,4 @@
 #define CSCS_APPEARANCE            0x1814
 
 void ble_init(void);
-void ble_notify_new_data(uint32_t, uint16_t, uint32_t, uint16_t);
+void ble_notify_new_data(uint32_t, uint16_t, uint16_t, uint16_t);
